Remove requiredness from demoDefaults.template
the manifest spec is derived from this file:
https://github.com/Financial-Times/ft-origami/blob/gh-pages/docs/syntax/origamijson.md

In that file, demosDefaults.template is marked as required, but in the example
it is not included.  I believe it was a mistake caused by copying and pasting
the options from the demos field.

In reality very few of our components include a default option for template, and
none of our tooling has ever relied on them doing so.

It's certainly strange, demosDefaults being optional but once included it must
list a default template??